requirements.txt
alembic==1.14.1
blinker==1.8.2
certifi==2023.11.17
click==8.1.7
cmake==3.30.4
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
distlib==0.3.7
dlib==19.24.6
dnspython==2.7.0
email_validator==2.2.0
et_xmlfile==2.0.0
face-recognition==1.3.0
face_recognition_models==0.3.0
filelock==3.13.1
Flask==3.0.3
Flask-Cors==4.0.1
flask-form==1.0.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-MySQL==1.5.2
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
Flask-WTF==1.2.2
fonttools==4.53.1
greenlet==3.0.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
Mako==1.3.8
MarkupSafe==2.1.5
matplotlib==3.9.2
mysql-connector-python==9.0.0
mysqlclient==2.2.4
numpy==2.1.1
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pillow==10.4.0
pipenv==2023.11.15
platformdirs==4.0.0
PyMySQL==1.1.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
setuptools==69.0.2
six==1.16.0
SQLAlchemy==2.0.31
typing_extensions==4.12.2
virtualenv==20.24.7
Werkzeug==3.0.3
WTForms==3.2.1